+++ docs/docs/changelog.md
@@ -25,6 +25,9 @@
   - 新增 `bind:pickend` 事件，当滚动选择结束时候触发事件 (by [@yawuling](https://github.com/yawuling) )
 - Input
   - 添加自定义外部类名 `custom-textarea-container-class` (by [@yawuling](https://github.com/yawuling) )
+- InputNumber
+  - 新增 `allow-null` 属性允许空值 (by [@yawuling](https://github.com/yawuling) )
+  - 新增 `placeholder` 属性设置占位文本 (by [@yawuling](https://github.com/yawuling) )
 - ImgCropper
   - 新增组件图片裁剪 (by [@HXCStudio123](https://github.com/HXCStudio123) )
 - MessageBox